refactor: separate for clarity
